feat(osdk): add LoongArch64 linker script template in base crate
Since the existing kernel entry point 0xFFFFFFFF80000000`` will be regarded as the physical address 0x8FFF80000000.
Therefore, we must manually specify that the ENTRY is the physical address 0x80000000, and explicitly KEEP the .text.entry to prevent the linker from discarding the entry function _start, and thus discarding the entire .text segment.