feat(ostd/la): add LoongArch boot code
Since LoongArch seems not to support large pages larger than 2MB (even if we map a 1GB large page, when the offset exceeds 0x200000, access to the space seems to return to the start of the mapped physical address, showing a roll - back phenomenon).
Therefore, we use 512 2MB large pages of the third - level page table to map the entire physical memory space (excluding the memory before 0x80000000). For the sake of memory conservation, we do not handle the address space before 0x80000000, because before initializing and activating the kernel page table, we only use linear mapping when processing the FDT. For this, we can temporarily use DMW1 to avoid the overhead of two page tables.