feat(ostd/la): introduce HUGE_MARKER bit to better distinguish huge pages
LoongArch defines bit 6 as the Global flag for common pages, but for huge pages, bit 6 is always set and bit 12 becomes the Global flag instead.

To distinguish huge pages in software without ambiguity, we introduce a software-defined HUGE_MARKER bit (bit 58), which is ignored by the MMU. This allows correct interpretation of Global flags depending on page type.